move user service to profile module
