+++ src/app/auth/auth.service.ts
@@ -9,8 +9,8 @@ import { from, Observable, of, Subscription } from 'rxjs';
 import { map, switchAll, switchMap, take, tap } from 'rxjs/operators';
 import { BaseService } from '../core/base.service';
 import { AlertService, Level, UntranslatedAlertMessage } from '../messaging/alert.service';
+import { User } from '../profile/user';
 import { LoginResult } from './login-result';
-import { User } from './user';
 
 export const LOGIN_URL = '/auth/login';
 export const LOGGED_OUT_URL = '/auth/logged-out';
