+++ src/app/individual/individual-detail/individual-subscription/individual-subscription.component.ts
@@ -2,8 +2,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { combineLatest, Observable, ReplaySubject } from 'rxjs';
 import { first, map } from 'rxjs/operators';
 import { AuthService } from '../../../auth/auth.service';
-import { UserService } from '../../../auth/user.service';
 import { AlertService } from '../../../messaging/alert.service';
+import { UserService } from '../../../profile/user.service';
 import { Individual } from '../../individual';
 
 @Component({
