+++ src/app/profile/profile-edit/profile-edit.component.ts
@@ -6,16 +6,16 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { first } from 'rxjs/operators';
 import { AuthService } from '../../auth/auth.service';
-import { User } from '../../auth/user';
-import { UserService } from '../../auth/user.service';
 import { BaseDetailComponent } from '../../core/base-detail.component';
 import { LanguageService } from '../../core/language.service';
 import { NavService } from '../../core/nav/nav.service';
 import { PublicUserService } from '../../open/public-user.service';
-import { ChangeEmailData } from '../change-email-dialog/change-email-data';
-import { ChangeEmailDialogComponent } from '../change-email-dialog/change-email-dialog.component';
-import { ChangePasswordData } from '../change-password-dialog/change-password-data';
-import { ChangePasswordDialogComponent } from '../change-password-dialog/change-password-dialog.component';
+import { User } from '../../profile/user';
+import { UserService } from '../../profile/user.service';
+import { ChangeEmailData } from './change-email-dialog/change-email-data';
+import { ChangeEmailDialogComponent } from './change-email-dialog/change-email-dialog.component';
+import { ChangePasswordData } from './change-password-dialog/change-password-data';
+import { ChangePasswordDialogComponent } from './change-password-dialog/change-password-dialog.component';
 
 @Component({
   templateUrl: './profile-edit.component.html',
