+++ src/app/profile/user.service.ts
@@ -1,16 +1,25 @@
 import { Injectable } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/firestore';
 import firebase from 'firebase/app';
-import { from, Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { combineLatest, from, Observable } from 'rxjs';
+import { filter, map, switchMap } from 'rxjs/operators';
 import { AuthService } from '../auth/auth.service';
 import { BaseResourceService } from '../core/base-resource.service';
+import { Individual } from '../individual/individual';
+import { IndividualService } from '../individual/individual.service';
+import { MasterdataService } from '../masterdata/masterdata.service';
 import { AlertService } from '../messaging/alert.service';
 import { User } from './user';
 
 @Injectable()
 export class UserService extends BaseResourceService<User> {
-  constructor(alertService: AlertService, protected afs: AngularFirestore, private authService: AuthService) {
+  constructor(
+    alertService: AlertService,
+    protected afs: AngularFirestore,
+    private authService: AuthService,
+    private individualService: IndividualService,
+    private masterdataService: MasterdataService
+  ) {
     super(alertService, afs, 'users');
   }
 
@@ -36,5 +45,15 @@ export class UserService extends BaseResourceService<User> {
 
   private followUnfollow(partial: Partial<unknown>): Observable<void> {
     return from(this.afs.collection('users').doc(this.authService.getUserId()).update(partial));
+  }
+
+  getFollowedIndividuals(limit$: Observable<number>): Observable<Individual[]> {
+    return combineLatest([this.authService.user$, limit$, this.masterdataService.phenoYear$]).pipe(
+      filter(
+        ([user, limit, year]) =>
+          user.following_individuals !== undefined && user.following_individuals.length !== 0 && year !== undefined
+      ),
+      switchMap(([user, limit, year]) => this.individualService.listByIds(user.following_individuals, year, limit))
+    );
   }
 }
